refactor(prediction): Fetch recent bets
