+++ src/views/Predictions/hooks/usePollPredictions.ts
@@ -2,7 +2,7 @@ import { useEffect, useRef } from 'react'
 import { useWeb3React } from '@web3-react/core'
 import { useAppDispatch } from 'state'
 import { useGetCurrentEpoch, useGetPredictionsStatus } from 'state/hooks'
-import { fetchMarketData, fetchRounds } from 'state/predictions'
+import { fetchLedgerData, fetchMarketData, fetchRounds } from 'state/predictions'
 import { PredictionStatus } from 'state/types'
 
 const POLL_TIME_IN_SECONDS = 10
@@ -25,6 +25,10 @@ const usePollPredictions = () => {
         const liveAndCurrent = [currentEpoch, currentEpoch - 1]
         dispatch(fetchRounds(liveAndCurrent))
         dispatch(fetchMarketData())
+
+        if (account) {
+          dispatch(fetchLedgerData({ account, epochs: liveAndCurrent }))
+        }
       }, POLL_TIME_IN_SECONDS * 1000)
     }
